Time: 26 ms (71.40%) | Memory: 52.6 MB (57.03%) - LeetSync

+++ 1014-k-closest-points-to-origin/k-closest-points-to-origin.java
@@ -11,11 +11,10 @@ class Solution {
         int[][] results = new int[k][];
         PriorityQueue<Point> pq = new PriorityQueue<>(Comparator.comparingDouble(point -> point.distance));
         for(int[] point : points){
-            // Omitting the square root calculation, as it's unnecessary for distance ranking
-            double distance = (long) point[0] * point[0] + (long) point[1] * point[1];
+            double distance = (double) point[0] * point[0] + (double) point[1] * point [1];
             pq.add(new Point(distance, point));
         }
-        for(int i=0;i<k;i++){
+        for(int i=0; i<k; i++){
             results[i] = pq.poll().point;
         }
         return results;
